requirements.txt
pymongo
google-generativeai
faiss-cpu
numpy
